fixed incorrect cast of metric values to int

+++ Products/ZenUtils/DaemonStats.py
@@ -1,10 +1,10 @@
 ##############################################################################
-# 
+#
 # Copyright (C) Zenoss, Inc. 2007, all rights reserved.
-# 
+#
 # This content is made available according to terms specified in
 # License.zenoss under the directory where your Zenoss product is installed.
-# 
+#
 ##############################################################################
 
 
@@ -100,10 +100,9 @@ class DaemonStats(object):
                 metric_min = 'U'
 
             value = self._derivative_tracker.derivative(
-                context_id, (int(value), timestamp),
+                context_id, (float(value), timestamp),
                 min=metric_min)
 
         # check for threshold breaches and send events when needed
         self._threshold_notifier.notify(
             self._contextKey(), context_id, self.name+'_'+name, timestamp, value)
-
